added FrameManager#prior_binding
Stores the binding prior to swapping in the FrameManager; this enables easy restore later.
Note this is only useful in the case where there was no active frame manager before swapping in the new one. In the case where
another frame manager takes control, then Framemanager#fresh_frame is used instead and FrameManager#prior_binding is not used or needed.